random.c: Expand RNG period to 2**64 - 1
This hopefully fixes the discrepancy between the measurements in the
Döbbelin, Schütt, Reinefeld paper and those generated by
test/qualitytest.c (cf. doc/qualitytest-8-8-8.txt).

+++ index.h
@@ -181,7 +181,7 @@ eqclass_from_index(const struct index_aux *aux, const struct index *idx)
 }
 
 /* random.c */
-extern unsigned random_seed;
+extern unsigned long long random_seed;
 extern void	random_index(const struct index_aux *, struct index *);
 
 /* cindex.c */
